// Source/WebSPA/ClientApp/src/app/shared/enums/appRoutes.ts
export enum AppRoutes {
    Equipments ='equipments',
    Employees = 'employees',
    Identity = 'identity',
    Softwares = 'softwares',
    IdentityConfirmEmail = 'identity/confirmEmail',
    IdentityChangePasswordAfterReset = 'identity/changePasswordAfterReset',
    IdentityChangePassword = 'identity/changePassword',
    Users = 'users',
    Administration = 'administration',
    Profile = 'profile'
}